add reserved metadata, reorg exerience vars/types
